Modify members in hmap from uint32_t to uint16_t typedef struct { const uint16_t len; const uint16_t data; } hcode16_t;

+++ libfaac/coder.h
@@ -86,8 +86,8 @@ typedef struct {
 #define DATASIZE (3*FRAME_LEN/2)
 
     struct {
-        int data;
-        int len;
+        short data;
+        short len;
     } s[DATASIZE];
     int datacnt;
